+++ src/common/rotate.glsl
@@ -3,3 +3,30 @@ mat2 rotate(float theta) {
   float s = sin(theta);
   return mat2(c, -s, s, c);    
 }
+
+mat3 rotateX(float theta) {
+  float c = cos(theta);
+  float s = sin(theta);
+  
+  return mat3(1, 0, 0, 
+              0, c, -s, 
+              0, s, c);
+}
+
+mat3 rotateY(float theta) {
+  float c = cos(theta);
+  float s = sin(theta);
+
+  return mat3(c, 0, s,
+              0, 1, 0,
+              -s, 0, c);
+}
+
+mat3 rotateZ(float theta) {
+  float c = cos(theta);
+  float s = sin(theta);
+
+  return mat3(c, -s, 0,
+              s, c, 0,
+              0, 0, 1);	    
+}