+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## **[1.4.0]** - 04/10/2020
 ### **Added**
 - Capsule sdf function
+- Torus sdf function
 
 ## **[1.4.0]** - 28/09/2020
 ### **Added**